+++ components/modal/symbol.ts
@@ -1,0 +1,9 @@
+export interface PositionPayload {
+  top: number,
+  left: number
+}
+
+export interface SizePayload {
+  width: number,
+  height: number
+}