+++ components/native-scroll/native-scroll.tsx
@@ -19,8 +19,10 @@ import {
   createEventEmitter,
   flatVNodes,
   isClient,
+  isDefined,
   isElement,
-  isTrue
+  isTrue,
+  warnOnce
 } from '@vexip-ui/utils'
 import { nativeScrollProps } from './props'
 import { useScrollWrapper } from './hooks'
@@ -35,13 +37,10 @@ const UP_EVENT = 'mouseup'
 
 export default defineComponent({
   name: 'NativeScroll',
-  components: {
-    Scrollbar,
-    ResizeObserver
-  },
+  inheritAttrs: false,
   props: nativeScrollProps,
   emits: [],
-  setup(_props, { slots, expose }) {
+  setup(_props, { attrs, slots, expose }) {
     const props = useProps('nativeScroll', _props, {
       scrollClass: null,
       scrollStyle: null,
@@ -76,8 +75,16 @@ export default defineComponent({
       barDuration: null,
       useBarTrack: false,
       scrollTag: 'div',
-      observeDeep: false
+      observeDeep: false,
+      scrollOnly: false
     })
+
+    if (isDefined(props.onBeforeScroll)) {
+      warnOnce(
+        "[vexip-ui:Table] 'on-before-scroll' prop has been deprecated, please " +
+          'do not use it anymore'
+      )
+    }
 
     const emitter = createEventEmitter()
 
@@ -389,13 +396,6 @@ export default defineComponent({
       event.stopPropagation()
       event.preventDefault()
 
-      const signX = event.clientX - cursorXPosition > 0 ? 1 : -1
-      const signY = event.clientY - cursorYPosition > 0 ? 1 : -1
-
-      if (props.onBeforeScroll?.({ signX, signY }) === false) {
-        return false
-      }
-
       scrolling.value = true
 
       if (enableXScroll.value) {
@@ -433,10 +433,7 @@ export default defineComponent({
 
       emitEvent(props.onWheel, event, type)
 
-      if (
-        (isVerticalScroll || isHorizontalScroll) &&
-        props.onBeforeScroll?.({ signX: sign, signY: sign }) !== false
-      ) {
+      if (isVerticalScroll || isHorizontalScroll) {
         const maxLimit = isVerticalScroll ? yScrollLimit.value : xScrollLimit.value
         const scroll = isVerticalScroll ? y.value : x.value
 
@@ -452,19 +449,8 @@ export default defineComponent({
       if (!contentElement.value) return
 
       event.stopPropagation()
-      event.preventDefault()
 
       const type = contentElement.value?.scrollLeft !== x.value ? 'horizontal' : 'vertical'
-
-      const signX = contentElement.value.scrollLeft - x.value > 0 ? 1 : -1
-      const signY = contentElement.value.scrollTop - y.value > 0 ? 1 : -1
-
-      if (props.onBeforeScroll?.({ signX, signY }) === false) {
-        contentElement.value.scrollTop = y.value
-        contentElement.value.scrollLeft = x.value
-
-        return
-      }
 
       y.value = contentElement.value.scrollTop
       x.value = contentElement.value.scrollLeft
@@ -610,7 +596,7 @@ export default defineComponent({
     }
 
     function renderContent() {
-      const Content = (props.scrollTag || 'div') as 'div'
+      const Content = (props.scrollTag || 'div') as any
       const children =
         props.observeDeep && slots.default ? renderSlot(slots, 'default', slotParams).children : []
 
@@ -618,8 +604,13 @@ export default defineComponent({
         <Content
           ref={contentElement}
           {...props.scrollAttrs}
+          {...(props.scrollOnly ? attrs : {})}
           class={wrapperClass.value}
-          style={[props.scrollAttrs?.style, props.scrollStyle]}
+          style={[props.scrollAttrs?.style, props.scrollStyle, props.scrollOnly && style.value]}
+          onMousedown={handleMouseDown}
+          onWheelPassive={(event: WheelEvent) =>
+            handleWheel(event, event.shiftKey ? 'horizontal' : 'vertical')
+          }
           onScroll={handleScroll}
         >
           {slots.extra && (
@@ -652,14 +643,12 @@ export default defineComponent({
     }
 
     return () => {
+      if (props.scrollOnly) {
+        return <ResizeObserver on-resize={handleResize}>{renderContent()}</ResizeObserver>
+      }
+
       return (
-        <div
-          ref={wrapper}
-          class={className.value}
-          style={style.value}
-          onMousedown={handleMouseDown}
-          onWheel={event => handleWheel(event, event.shiftKey ? 'horizontal' : 'vertical')}
-        >
+        <div ref={wrapper} {...attrs} class={className.value} style={style.value}>
           <ResizeObserver on-resize={handleResize}>{renderContent()}</ResizeObserver>
           {props.useXBar && (
             <Scrollbar