+++ components/native-scroll/props.ts
@@ -34,6 +34,7 @@ export const nativeScrollProps = buildProps({
   useBarTrack: booleanProp,
   scrollTag: String,
   observeDeep: booleanProp,
+  scrollOnly: booleanProp,
   onResize: eventProp<(entry: ResizeObserverEntry) => void>(),
   onXEnabledChange: eventProp<(enabled: boolean) => void>(),
   onYEnabledChange: eventProp<(enabled: boolean) => void>(),