+++ components/textarea/props.ts
@@ -1,4 +1,11 @@
-import { booleanProp, buildProps, eventProp, localeProp, stateProp } from '@vexip-ui/config'
+import {
+  booleanProp,
+  buildProps,
+  classProp,
+  eventProp,
+  localeProp,
+  stateProp
+} from '@vexip-ui/config'
 
 import type { ExtractPropTypes, PropType } from 'vue'
 import type { IconEffect } from '@/components/icon'
@@ -17,12 +24,14 @@ export const textareaProps = buildProps({
   readonly: booleanProp,
   disabled: booleanProp,
   debounce: booleanProp,
+  delay: Number,
   maxLength: Number,
   loading: booleanProp,
   loadingIcon: Object,
   loadingLock: booleanProp,
   loadingEffect: String as PropType<IconEffect>,
   sync: booleanProp,
+  controlClass: classProp,
   onFocus: eventProp<(event: FocusEvent) => void>(),
   onBlur: eventProp<(event: FocusEvent) => void>(),
   onInput: eventProp<(value: string) => void>(),